requirements.txt
fastapi==0.79.0
uvicorn==0.18.2
Pillow==9.2.0
flake8==5.0.2
pytest==7.1.2
